llm-gateway/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==1.10.13
httpx==0.27.0
pyjwt==2.8.0